refactor: split userStore into domain slices behind unchanged public API
