+++ src/store/slices/uiSlice.ts
@@ -1,0 +1,6 @@
+// UI request state (isLoading/error). Written by other slices via the shared set().
+export const createUiSlice = () => ({
+  // Initial state
+  isLoading: false,
+  error: null,
+});
